libretro-dolphin: import version 20180827
Libretro is a simple but powerful development interface that allows for
the easy creation of emulators, games and multimedia applications that
can plug straight into any libretro-compatible frontend. This development
interface is open to others so that they can run these pluggable emulator
and game cores also in their own programs or devices.

Dolphin is a Nintendo GameCube and Wii emulator with high compatibility
across the majority of titles for both platforms.