Link PHP with shared libsqlite3 isntead of built-in
We used to build PHP with its built-in, statically linked libsqlite3. When
used in an executable with dynamically looaded modules such as Apache, some
module may load a shared libsqlite3, which has the same symbols as PHP's
built-in libsqlite3. This causes unreliable cross-version calls and is
source of crashes.

The fix is to disable PHP's built-in libslite3 and always use an external,
shared libsqlite3.